- Add support for Dynamic Frameworks and Carthage - Convert BVSDKDemo to Swift 3 syntax - Cleanup compiler warnings - Remove deprecated Conversations API from BVSDK 2.x - Update Gimbal.framework to 2.52.1 - Fix use of %lu and %li in some spots - Fix missing negation operator when checking for array type in the BVRatingDistribution - Serveral updates for ci scripts and matrixed travis ci builds

+++ Pod/BVCommon/BVViewsHelper.m
@@ -42,7 +42,7 @@
 
 
 +(NSString*)formatIndex:(NSIndexPath*)indexPath {
-    return [NSString stringWithFormat:@"%li:%li", indexPath.section, indexPath.row];
+    return [NSString stringWithFormat:@"%lu:%lu", (unsigned long)indexPath.section, (unsigned long)indexPath.row];
 }
 
 @end
