+++ Pod/BVConversations/Submission/Question/BVQuestionSubmission.m
@@ -148,7 +148,7 @@
             NSDictionary *json = [NSJSONSerialization JSONObjectWithData:data options:kNilOptions error:&jsonParsingError];
             BVQuestionSubmissionErrorResponse* errorResponse = [[BVQuestionSubmissionErrorResponse alloc] initWithApiResponse:json]; // fails gracefully
             
-            [[BVLogger sharedLogger] verbose:[NSString stringWithFormat:@"RESPONSE: %@ (%d)", json, statusCode]];
+            [[BVLogger sharedLogger] verbose:[NSString stringWithFormat:@"RESPONSE: %@ (%ld)", json, (long)statusCode]];
             
             if (httpError) {
                 // network error was generated
